+++ client/app/bundles/course/survey/constants.js
@@ -76,7 +76,6 @@ const actionTypes = mirrorCreator([
   'QUESTION_FORM_HIDE',
   'SECTION_FORM_SHOW',
   'SECTION_FORM_HIDE',
-  'SET_DRAGGED_QUESTION',
   'REORDER_QUESTION',
   'CHANGE_QUESTION_SECTION',
   'UPDATE_QUESTION_ORDER_REQUEST',
